refactor: extract reusable components from SearchView
- Extract SearchIconButton into separate component
- Extract TruncatedText into separate component
- Extract ReverseImageSearchButton into separate component
- Extract ImageWithBlurhash into separate component
- Extract VideoWithBlurhash into separate component
- Extract RelayStatusDisplay into separate component
- Update SearchView to use extracted components
- Reduce SearchView.tsx from 2577 to ~2000 lines
- Improve code reusability and maintainability
