+++ src/components/SearchView.tsx
@@ -10,7 +10,6 @@ import { applyContentFilters, isEmojiSearch } from '@/lib/contentAnalysis';
 import { isAbsoluteHttpUrl, formatUrlForDisplay, extractImageUrls, extractVideoUrls, extractNonMediaUrls, getFilenameFromUrl } from '@/lib/utils/urlUtils';
 import { updateSearchQuery } from '@/lib/utils/navigationUtils';
 import { extractImetaImageUrls, extractImetaVideoUrls, extractImetaBlurhashes, extractImetaDimensions, extractImetaHashes } from '@/lib/picture';
-import { Blurhash } from 'react-blurhash';
 // Use unified cached NIP-05 checker for DRYness and to leverage persistent cache
 import { checkNip05 as verifyNip05Async } from '@/lib/vertex';
 
@@ -18,7 +17,6 @@ import { useSearchParams, useRouter, usePathname } from 'next/navigation';
 import { getCurrentProfileNpub, toImplicitUrlQuery, toExplicitInputFromUrl, ensureAuthorForBackend, decodeUrlQuery } from '@/lib/search/queryTransforms';
 import { profileEventFromPubkey } from '@/lib/vertex';
 import { getProfileScopeIdentifiers, hasProfileScope, addProfileScope, removeProfileScope } from '@/lib/search/profileScope';
-import Image from 'next/image';
 import EventCard from '@/components/EventCard';
 import UrlPreview from '@/components/UrlPreview';
 import ProfileCard from '@/components/ProfileCard';
@@ -27,6 +25,10 @@ import CopyButton from '@/components/CopyButton';
 import ProfileScopeIndicator from '@/components/ProfileScopeIndicator';
 import FilterCollapsed from '@/components/FilterCollapsed';
 import RelayCollapsed from '@/components/RelayCollapsed';
+import RelayStatusDisplay from '@/components/RelayStatusDisplay';
+import TruncatedText from '@/components/TruncatedText';
+import ImageWithBlurhash from '@/components/ImageWithBlurhash';
+import VideoWithBlurhash from '@/components/VideoWithBlurhash';
 import { nip19 } from 'nostr-tools';
 import { extractNip19Identifiers, decodeNip19Pointer } from '@/lib/utils/nostrIdentifiers';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -40,137 +42,9 @@ import { formatEventTimestamp } from '@/lib/utils/eventHelpers';
 import { TEXT_MAX_LENGTH, TEXT_LINK_CHAR_COUNT, SEARCH_FILTER_THRESHOLD } from '@/lib/constants';
 import { HIGHLIGHTS_KIND } from '@/lib/highlights';
 
-// Reusable search icon button component
-function SearchIconButton({ 
-  onClick, 
-  title, 
-  className = "" 
-}: { 
-  onClick: () => void; 
-  title: string; 
-  className?: string; 
-}) {
-  return (
-    <button
-      type="button"
-      className={`absolute top-1.5 right-1.5 z-10 w-7 h-7 flex items-center justify-center text-gray-500 hover:text-gray-300 bg-black/30 hover:bg-black/50 border border-gray-600/40 hover:border-gray-500/60 rounded-sm opacity-60 hover:opacity-100 transition-all duration-200 ${className}`}
-      title={title}
-      onClick={(e) => {
-        e.stopPropagation();
-        onClick();
-      }}
-    >
-      <FontAwesomeIcon icon={faMagnifyingGlass} className="w-3 h-3" />
-    </button>
-  );
-}
 
 
-// Component for truncating long text with fade effect and expand arrow
-function TruncatedText({ 
-  content, 
-  maxLength = TEXT_MAX_LENGTH, 
-  className = '',
-  renderContentWithClickableHashtags
-}: { 
-  content: string; 
-  maxLength?: number; 
-  className?: string;
-  renderContentWithClickableHashtags: (content: string) => React.ReactNode;
-}) {
-  const [isExpanded, setIsExpanded] = useState(false);
-  
-  if (!content) return null;
-  
-  // Calculate effective length considering links as 10 characters each
-  const calculateEffectiveLength = (text: string): number => {
-    // Regex patterns for different types of links
-    const urlPattern = /https?:\/\/[^\s]+/g;
 
-    let effectiveLength = text.length;
-
-    const urls = text.match(urlPattern) || [];
-    urls.forEach(url => {
-      effectiveLength = effectiveLength - url.length + TEXT_LINK_CHAR_COUNT;
-
-      const nestedIdentifiers = extractNip19Identifiers(url);
-      nestedIdentifiers.forEach(identifier => {
-        effectiveLength = effectiveLength - identifier.length + TEXT_LINK_CHAR_COUNT;
-      });
-    });
-
-    const directIdentifiers = extractNip19Identifiers(text);
-    directIdentifiers.forEach(identifier => {
-      const alreadyCovered = urls.some(url => url.includes(identifier));
-      if (alreadyCovered) return;
-      effectiveLength = effectiveLength - identifier.length + TEXT_LINK_CHAR_COUNT;
-    });
-
-    return effectiveLength;
-  };
-  
-  const effectiveLength = calculateEffectiveLength(content);
-  const shouldTruncate = effectiveLength > maxLength;
-  
-  // For display, we still need to truncate the actual text
-  const displayText = isExpanded || !shouldTruncate ? content : content.slice(0, maxLength);
-  
-  return (
-    <div className={`relative ${className}`}>
-      <div className={shouldTruncate && !isExpanded ? 'relative' : ''}>
-        {renderContentWithClickableHashtags(displayText)}
-        {shouldTruncate && !isExpanded && (
-          <div className="absolute bottom-0 left-0 right-0 h-16 bg-gradient-to-t from-[#2d2d2d] to-transparent pointer-events-none" />
-        )}
-      </div>
-      {shouldTruncate && (
-        <div className="mt-0.5">
-          <button
-            type="button"
-            onClick={() => setIsExpanded(!isExpanded)}
-            className="w-full flex items-center justify-center text-gray-400 hover:text-gray-200 transition-colors"
-          >
-            <FontAwesomeIcon 
-              icon={isExpanded ? faChevronUp : faChevronDown} 
-              className="w-3 h-3" 
-            />
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Reusable reverse image search button component
-function ReverseImageSearchButton({ 
-  imageUrl, 
-  className = "" 
-}: { 
-  imageUrl: string; 
-  className?: string; 
-}) {
-  const handleReverseSearch = () => {
-    const lensUrl = `https://lens.google.com/uploadbyurl?url=${encodeURIComponent(imageUrl)}`;
-    window.open(lensUrl, '_blank', 'noopener,noreferrer');
-  };
-
-  return (
-    <button
-      type="button"
-      className={`absolute top-1.5 left-1.5 z-10 w-7 h-7 flex items-center justify-center text-gray-500 hover:text-gray-300 bg-black/30 hover:bg-black/50 border border-gray-600/40 hover:border-gray-500/60 rounded-sm opacity-60 hover:opacity-100 transition-all duration-200 ${className}`}
-      title="Reverse image search with Google Lens (external)"
-      onClick={(e) => {
-        e.stopPropagation();
-        handleReverseSearch();
-      }}
-    >
-      <div className="relative">
-        <FontAwesomeIcon icon={faEye} className="w-3 h-3" />
-        <FontAwesomeIcon icon={faExternalLink} className="absolute -top-1 -right-1 w-2 h-2 text-gray-400" />
-      </div>
-    </button>
-  );
-}
 // Removed direct Highlight usage; RawEventJson handles JSON highlighting
 // import { Highlight, themes, type RenderProps } from 'prism-react-renderer';
 import RawEventJson from '@/components/RawEventJson';
@@ -185,287 +59,7 @@ type Props = {
   onUrlUpdate?: (query: string) => void;
 };
 
-// Component to handle image loading with blurhash placeholder
-function ImageWithBlurhash({ 
-  src, 
-  blurhash, 
-  alt, 
-  width, 
-  height, 
-  dim,
-  onClickSearch
-}: {
-  src: string;
-  blurhash?: string;
-  alt: string;
-  width: number;
-  height: number;
-  dim?: { width: number; height: number } | null;
-  onClickSearch?: () => void;
-}) {
-  const [imageLoaded, setImageLoaded] = useState(false);
-  const [imageError, setImageError] = useState(false);
-  const [statusCode, setStatusCode] = useState<number | null>(null);
-  const [measuredDim, setMeasuredDim] = useState<{ width: number; height: number } | null>(null);
 
-  useEffect(() => {
-    setImageLoaded(false);
-    setImageError(false);
-    setMeasuredDim(null);
-  }, [src]);
-
-  if (!isAbsoluteHttpUrl(src)) {
-    return null;
-  }
-
-  const effectiveDim = dim && dim.width > 0 && dim.height > 0 ? dim : measuredDim;
-  const aspectStyle = effectiveDim
-    ? { aspectRatio: `${effectiveDim.width} / ${effectiveDim.height}` }
-    : { minHeight: '200px' as const };
-
-  return (
-    <div 
-      className="relative w-full overflow-hidden rounded-md border border-[#3d3d3d] bg-[#1f1f1f] group"
-      style={aspectStyle}
-    >
-      {/* Blurhash placeholder - shown while loading or on error */}
-      {blurhash && (!imageLoaded || imageError) && (
-        <div className="absolute inset-0">
-          <Blurhash 
-            hash={blurhash} 
-            width={'100%'} 
-            height={'100%'} 
-            resolutionX={32} 
-            resolutionY={32} 
-            punch={1} 
-          />
-        </div>
-      )}
-
-      {/* Subtle loading spinner on top of blurhash while the image loads */}
-      {!imageLoaded && !imageError && (
-        <div className="absolute inset-0 flex items-center justify-center z-10">
-          <div
-            className="h-6 w-6 rounded-full border-2 border-gray-300/70 border-t-transparent animate-spin"
-            aria-label="Loading image"
-          />
-        </div>
-      )}
-
-      {/* Error state: show status code while keeping blurhash (if any) */}
-      {imageError && (
-        <div className="absolute inset-0 flex items-center justify-center z-10">
-          <div className="px-3 py-2 rounded-md bg-black/40 text-gray-200 text-sm flex items-center justify-center gap-2 border border-[#3d3d3d]">
-            <FontAwesomeIcon icon={faImage} className="opacity-80" />
-            <span className="flex-1 text-center">{statusCode ?? 'Error'}</span>
-            <button
-              type="button"
-              className="p-1 hover:bg-white/10 rounded transition-colors"
-              title="Open image in new tab"
-              onClick={(e) => {
-                e.stopPropagation();
-                window.open(src, '_blank', 'noopener,noreferrer');
-              }}
-            >
-              <FontAwesomeIcon icon={faExternalLink} className="text-xs opacity-80" />
-            </button>
-          </div>
-        </div>
-      )}
-      
-      {/* Real image - hidden until loaded */}
-      <Image 
-        src={trimImageUrl(src)} 
-        alt={alt}
-        width={width}
-        height={height} 
-        className={`absolute inset-0 h-full w-full object-contain transition-opacity duration-300 ${
-          imageLoaded ? 'opacity-100' : 'opacity-0'
-        }`}
-        unoptimized
-        onLoad={(e) => { 
-          setImageLoaded(true); 
-          setStatusCode(200); 
-          try {
-            const img = e.currentTarget as HTMLImageElement;
-            if (!effectiveDim && img?.naturalWidth && img?.naturalHeight) {
-              setMeasuredDim({ width: img.naturalWidth, height: img.naturalHeight });
-            }
-          } catch {}
-        }}
-        onError={() => {
-          setImageError(true);
-          try {
-            // Some browsers expose a 'naturalWidth' of 0 on 404 but no status code; try fetch HEAD
-            fetch(src, { method: 'HEAD' }).then((res) => {
-              setStatusCode(res.status || null);
-            }).catch(() => setStatusCode(null));
-          } catch { setStatusCode(null); }
-        }}
-      />
-      
-      {/* Search icon button - only show when image is loaded and onClickSearch is provided */}
-      {imageLoaded && !imageError && onClickSearch && (
-        <SearchIconButton
-          onClick={onClickSearch}
-          title="Search for this image"
-        />
-      )}
-      
-      {/* Reverse image search button - only show when image is loaded */}
-      {imageLoaded && !imageError && (
-        <ReverseImageSearchButton
-          imageUrl={src}
-        />
-      )}
-
-      {/* Copy image URL button - bottom right */}
-      {imageLoaded && !imageError && (
-        <div className="absolute bottom-1.5 right-1.5 z-10">
-          <CopyButton 
-            text={src} 
-            title="Copy image URL" 
-            className="w-7 h-7 text-gray-500 hover:text-gray-300 bg-black/30 hover:bg-black/50 border border-gray-600/40 hover:border-gray-500/60 rounded-sm opacity-60 hover:opacity-100 transition-all duration-200"
-          />
-        </div>
-      )}
-
-      {/* Open external button - bottom left */}
-      {imageLoaded && !imageError && (
-        <button
-          type="button"
-          className="absolute bottom-1.5 left-1.5 z-10 w-7 h-7 flex items-center justify-center text-gray-500 hover:text-gray-300 bg-black/30 hover:bg-black/50 border border-gray-600/40 hover:border-gray-500/60 rounded-sm opacity-60 hover:opacity-100 transition-all duration-200"
-          title="Open image in new tab"
-          onClick={(e) => {
-            e.stopPropagation();
-            window.open(src, '_blank', 'noopener,noreferrer');
-          }}
-        >
-          <FontAwesomeIcon icon={faExternalLink} className="w-3 h-3" />
-        </button>
-      )}
-    </div>
-  );
-}
-
-// Component to handle video loading with blurhash placeholder
-function VideoWithBlurhash({ 
-  src, 
-  blurhash, 
-  dim,
-  onClickSearch
-}: {
-  src: string;
-  blurhash?: string;
-  dim?: { width: number; height: number } | null;
-  onClickSearch?: () => void;
-}) {
-  const [videoLoaded, setVideoLoaded] = useState(false);
-  const [videoError, setVideoError] = useState(false);
-  const [statusCode, setStatusCode] = useState<number | null>(null);
-
-  useEffect(() => {
-    setVideoLoaded(false);
-    setVideoError(false);
-  }, [src]);
-
-  if (!isAbsoluteHttpUrl(src)) {
-    return null;
-  }
-
-  const aspectStyle = dim && dim.width > 0 && dim.height > 0
-    ? { aspectRatio: `${dim.width} / ${dim.height}` }
-    : { minHeight: '200px' as const };
-
-  return (
-    <div 
-      className="relative w-full overflow-hidden rounded-md border border-[#3d3d3d] bg-[#1f1f1f] group"
-      style={aspectStyle}
-    >
-      {/* Blurhash placeholder - shown while loading or on error */}
-      {blurhash && (!videoLoaded || videoError) && (
-        <div className="absolute inset-0">
-          <Blurhash 
-            hash={blurhash} 
-            width={'100%'} 
-            height={'100%'} 
-            resolutionX={32} 
-            resolutionY={32} 
-            punch={1} 
-          />
-        </div>
-      )}
-
-      {/* Subtle loading spinner on top of blurhash while the video loads */}
-      {!videoLoaded && !videoError && (
-        <div className="absolute inset-0 flex items-center justify-center z-10">
-          <div
-            className="h-6 w-6 rounded-full border-2 border-gray-300/70 border-t-transparent animate-spin"
-            aria-label="Loading video"
-          />
-        </div>
-      )}
-
-      {/* Error state: show status code while keeping blurhash (if any) */}
-      {videoError && (
-        <div className="absolute inset-0 flex items-center justify-center z-10">
-          <div className="px-3 py-2 rounded-md bg-black/40 text-gray-200 text-sm flex items-center justify-center gap-2 border border-[#3d3d3d]">
-            <FontAwesomeIcon icon={faImage} className="opacity-80" />
-            <span className="flex-1 text-center">{statusCode ?? 'Error'}</span>
-            <button
-              type="button"
-              className="p-1 hover:bg-white/10 rounded transition-colors"
-              title="Open video in new tab"
-              onClick={(e) => {
-                e.stopPropagation();
-                window.open(src, '_blank', 'noopener,noreferrer');
-              }}
-            >
-              <FontAwesomeIcon icon={faExternalLink} className="text-xs opacity-80" />
-            </button>
-          </div>
-        </div>
-      )}
-      
-      {/* Real video - hidden until loaded */}
-      <video 
-        controls 
-        playsInline 
-        className={`absolute inset-0 h-full w-full object-cover transition-opacity duration-300 ${
-          videoLoaded ? 'opacity-100' : 'opacity-0'
-        }`}
-        onLoadedData={() => { setVideoLoaded(true); setStatusCode(200); }}
-        onError={() => {
-          setVideoError(true);
-          try {
-            // Try to fetch HEAD to get status code
-            fetch(src, { method: 'HEAD' }).then((res) => {
-              setStatusCode(res.status || null);
-            }).catch(() => setStatusCode(null));
-          } catch { setStatusCode(null); }
-        }}
-      >
-        <source src={src} />
-        Your browser does not support the video tag.
-      </video>
-      
-      {/* Search icon button - only show when video is loaded and onClickSearch is provided */}
-      {videoLoaded && !videoError && onClickSearch && (
-        <SearchIconButton
-          onClick={onClickSearch}
-          title="Search for this video"
-        />
-      )}
-      
-      {/* Reverse image search button - only show when video is loaded */}
-      {videoLoaded && !videoError && (
-        <ReverseImageSearchButton
-          imageUrl={src}
-        />
-      )}
-    </div>
-  );
-}
 
 // (Local AuthorBadge removed; using global `components/AuthorBadge` inside EventCard.)
 
@@ -2111,133 +1705,10 @@ export default function SearchView({ initialQuery = '', manageUrl = true, onUrlU
 
           {/* Expanded views - below button row, full width */}
           {showConnectionDetails && connectionDetails && (
-            <div className="mt-2 p-3 bg-[#2d2d2d] border border-[#3d3d3d] rounded-lg text-xs w-full">
-              
-              {/* Connected relays */}
-              {connectionDetails?.connectedRelays && connectionDetails.connectedRelays.length > 0 && (
-                <div className="mb-2">
-                  <div className="text-green-400 font-medium mb-1">
-                    ✅ Connected ({connectionDetails.connectedRelays.length})
-                  </div>
-                  <div className="space-y-1">
-                    {connectionDetails.connectedRelays.map((relay, idx) => {
-                      const ping = connectionDetails?.relayPings?.get(relay);
-                      const pingDisplay = ping && ping > 0 ? ` (${ping}ms)` : '';
-                      return (
-                        <div key={idx} className="text-gray-300 ml-2">
-                          • {relay.replace(/^wss:\/\//, '').replace(/\/$/, '')}{pingDisplay}
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-
-              {/* Recently active relays (but not currently connected or failed) */}
-              {(() => {
-                const connectedSet = new Set(connectionDetails?.connectedRelays || []);
-                const failedSet = new Set(connectionDetails?.failedRelays || []);
-                const connectingSet = new Set(connectionDetails?.connectingRelays || []);
-                
-                // Debug logging
-                console.log('Relay status debug:', {
-                  connected: connectionDetails?.connectedRelays || [],
-                  failed: connectionDetails?.failedRelays || [],
-                  connecting: connectionDetails?.connectingRelays || [],
-                  recentlyActive: recentlyActive
-                });
-                
-                // Only show recently active relays that are not connected, not failed, and not connecting
-                const recentlyActiveOnly = recentlyActive.filter(relay => 
-                  !connectedSet.has(relay) && 
-                  !failedSet.has(relay) && 
-                  !connectingSet.has(relay)
-                );
-                
-                if (recentlyActiveOnly.length === 0) return null;
-                return (
-                  <div className="mb-2">
-                    <div className="text-blue-400 font-medium mb-1">
-                      🔵 Recently active ({recentlyActiveOnly.length})
-                    </div>
-                    <div className="space-y-1">
-                      {recentlyActiveOnly.map((relay, idx) => {
-                        const ping = connectionDetails?.relayPings?.get(relay);
-                        const pingDisplay = ping && ping > 0 ? ` (${ping}ms)` : '';
-                        return (
-                          <div key={idx} className="text-gray-300 ml-2">
-                            • {relay.replace(/^wss:\/\//, '').replace(/\/$/, '')}{pingDisplay}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-                );
-              })()}
-              
-              {/* Connecting relays */}
-              {connectionDetails?.connectingRelays && connectionDetails.connectingRelays.length > 0 && (
-                <div className="mb-2">
-                  <div className="text-yellow-400 font-medium mb-1">
-                    🟡 Connecting ({connectionDetails.connectingRelays.length})
-                  </div>
-                  <div className="space-y-1">
-                    {connectionDetails.connectingRelays.map((relay, idx) => {
-                      const ping = connectionDetails?.relayPings?.get(relay);
-                      const pingDisplay = ping && ping > 0 ? ` (${ping}ms)` : '';
-                      return (
-                        <div key={idx} className="text-gray-300 ml-2">
-                          • {relay.replace(/^wss:\/\//, '').replace(/\/$/, '')}{pingDisplay}
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-              
-              {(() => {
-                const connectedSet = new Set(connectionDetails?.connectedRelays || []);
-                const recentlyActiveSet = new Set(recentlyActive);
-                const connectingSet = new Set(connectionDetails?.connectingRelays || []);
-                
-                // Only show relays that are truly failed (not connected, not recently active, not connecting)
-                const failedFiltered = (connectionDetails?.failedRelays || []).filter((relay) => 
-                  !connectedSet.has(relay) && 
-                  !recentlyActiveSet.has(relay) && 
-                  !connectingSet.has(relay)
-                );
-                
-                if (failedFiltered.length === 0) return null;
-                return (
-                  <div>
-                    <div className="text-red-400 font-medium mb-1">
-                      ❌ Failed ({failedFiltered.length})
-                    </div>
-                    <div className="space-y-1">
-                      {failedFiltered.map((relay, idx) => {
-                        const ping = connectionDetails?.relayPings?.get(relay);
-                        const pingDisplay = ping && ping > 0 ? ` (${ping}ms)` : '';
-                        return (
-                          <div key={idx} className="text-gray-300 ml-2">
-                            • {relay.replace(/^wss:\/\//, '').replace(/\/$/, '')}{pingDisplay}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-                );
-              })()}
-              
-              {(() => {
-                const anyReachable = (connectionDetails?.connectedRelays?.length || 0) > 0 || recentlyActive.length > 0;
-                const anyFailed = (connectionDetails?.failedRelays?.length || 0) > 0;
-                return (!anyReachable && !anyFailed) ? (
-                <div className="text-gray-400">
-                  No relay connection information available
-                </div>
-                ) : null;
-              })()}
-            </div>
+            <RelayStatusDisplay 
+              connectionDetails={connectionDetails}
+              recentlyActive={recentlyActive}
+            />
           )}
 
           {showFilterDetails && (
